tvl_backends/tvl-backends-nvdec/requirements.txt
numpy==1.14.6
Pillow==5.2.0
pytest==3.8.2
torch==1.0.0
